refactor(website): add about page
