+++ platform/website/src/routes/(app)/+layout.ts
