+++ platform/website/src/routes/(app)/[username]/+page.svelte
@@ -138,7 +138,7 @@
 </div>
 
 <style lang="scss">
-	@import "../../assets/styles/variables.scss";
+	@import "../../../assets/styles/variables.scss";
 
 	.content {
 		grid-area: content;
