+++ platform/website/src/routes/(app)/following/+page.svelte
@@ -49,7 +49,7 @@
 </div>
 
 <style lang="scss">
-	@import "../../assets/styles/variables.scss";
+	@import "../../../assets/styles/variables.scss";
 
 	.content {
 		grid-area: content;
