+++ platform/website/src/routes/(app)/search/+page.svelte
@@ -120,7 +120,7 @@
 {/if}
 
 <style lang="scss">
-	@import "../../assets/styles/variables.scss";
+	@import "../../../assets/styles/variables.scss";
 
 	.no-results {
 		color: $textColorLight;
