+++ platform/website/src/routes/(app)/search/+page.ts
