+++ platform/website/src/routes/(app)/settings/+layout.svelte
@@ -52,7 +52,7 @@
 </div>
 
 <style lang="scss">
-	@import "../../assets/styles/variables.scss";
+	@import "../../../assets/styles/variables.scss";
 
 	.content {
 		grid-area: content;
