+++ platform/website/src/routes/(app)/sign-up/+page.ts
