+++ platform/website/src/routes/(app)/tag/[tag]/+page.ts
