+++ platform/website/src/routes/about/+page.svelte
@@ -1,0 +1,5 @@
+<svelte:head>
+	<title>Scuffle</title>
+</svelte:head>
+
+<div>Troy is funny Kappa</div>
